refactor-fe: 리뷰 후 폴더명 수정
컴포넌트 폴더 안에 폴더 대문자 => 소문자
