+++ frontend/src/components/profile/ProfileImage.tsx
