+++ frontend/src/pages/camps/campInfo.tsx
@@ -2,10 +2,10 @@ import { useEffect, useState } from 'react';
 import { useSuspenseQuery } from '@tanstack/react-query';
 import { useParams } from 'react-router-dom';
 import CheckIcon from '../../assets/icons/checkIcon.svg?react';
-import SubscribeButton from '../../components/Button/SubscribeButton';
+import SubscribeButton from '../../components/button/SubscribeButton';
 import Text from '../../components/ui/Text';
 import { CampInfo as CampInfoType } from '../../types/api/camp';
-import ProfileImage from '../../components/ProfileImage/ProfileImage';
+import ProfileImage from '../../components/profile/ProfileImage';
 import useAuth from '../../hooks/useAuth';
 import useFetch from '../../hooks/useFetch';
 import { isSubscribedCamp, subscribeCamp } from '../../API/subscription';
